+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+
+
+## [0.1.3] - 2020-03-18
+### Changed
+    - Added documentation how to use config manager
+    - Secret manager config parameter
 
 ## [0.1.2] - 2020-03-17
 ### Added
@@ -29,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - MSSql runner classes
     - Batch control
 
+[0.1.3]: https://github.com/equinoxfitness/mssql-runner/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/equinoxfitness/mssql-runner/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/equinoxfitness/mssql-runner/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/equinoxfitness/mssql-runner/releases/tag/v0.1.0